+++ android/app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Libum"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.view.activity.MainActivity"
             android:exported="true"
             android:label="@string/app_name">
             <intent-filter>
