+++ android/app/src/main/java/com/project/libum/data/model/DashboardViewModel.kt
@@ -1,13 +1,13 @@
-package com.project.libum.ui.home
+package com.project.libum.data.model
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class DashboardViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is dashboard Fragment"
     }
     val text: LiveData<String> = _text
 }
