+++ android/app/src/main/java/com/project/libum/data/model/NotificationsViewModel.kt
@@ -1,13 +1,13 @@
-package com.project.libum.ui.dashboard
+package com.project.libum.data.model
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class NotificationsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is notifications Fragment"
     }
     val text: LiveData<String> = _text
 }
