+++ android/app/src/main/res/drawable/ic_bottom_bacground.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="390dp"
+    android:height="74dp"
+    android:viewportWidth="390"
+    android:viewportHeight="74">
+    <path
+        android:fillColor="#FFD9D9D9"
+        android:pathData="M390 16.66v57.27H0V9.34C34.75-2.58 74.97 2.27 115.84 7.2c25.82 3.11 51.9 6.26 77.03 5.23 19.53-0.8 38.89-3.53 58.03-6.23 44.13-6.2 87.12-12.25 128.4 5.78 3.53 1.53 7.1 3.1 10.7 4.68Z"/>
+</vector>
